+++ plugin/src/main/java/com/instana/android/plugin/IgnoredClasses.kt
@@ -13,7 +13,7 @@ object IgnoredClasses {
 
     val troublesomeAnalytics = listOf(
         Regex("com\\.appsflyer\\..*"),
-        Regex("com\\.google\\.android\\.gms\\.internal\\..*"),
+        Regex("com\\.google\\.android\\.gms\\.internal\\.(?!ads\\.).*"),
         Regex("com\\.google\\.firebase\\..*"),
         Regex("com\\.instabug\\.library\\..*"),
         Regex("com\\.newrelic\\..*"),